# cmd/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.CreateMaterialRequest:
    properties:
      isotopes:
        type: string
      material_description:
        type: string
      material_name:
        type: string
      sample_requirements:
        type: string
    required:
    - isotopes
    - material_name
    - sample_requirements
    type: object
...
  handler.UpdateMaterialRequest:
    properties:
      isotopes:
        type: string
      material_description:
        type: string
      material_name:
        type: string
      sample_requirements:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Бэкенд сервис для анализа материалов и радиоуглеродного датирования.
  title: Система анализа материалов
  version: "1.0"
paths:
...
  /materials:
    get:
      description: Get all materials with optional search
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response with materials array
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all materials
      tags:
      - Materials
    post:
      consumes:
      - application/json
      description: Create new material (admin only)
      parameters:
      - description: Material data
        in: body
        name: material
        required: true
        schema:
          $ref: '#/definitions/handler.CreateMaterialRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Material created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create material
      tags:
      - Materials
  /materials/{id}:
    delete:
      description: Delete material by ID (admin only)
      parameters:
      - description: Material ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Material deleted successfully
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete material
      tags:
      - Materials
    get:
      description: Get material details by ID
      parameters:
      - description: Material ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success response with material
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Material not found
          schema:
            additionalProperties: true
            type: object
      summary: Get material by ID
      tags:
      - Materials
    put:
      consumes:
      - application/json
      description: Update material by ID (admin only)
      parameters:
      - description: Material ID
        in: path
        name: id
        required: true
        type: integer
      - description: Material update data
        in: body
        name: material
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateMaterialRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Material updated successfully
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update material
      tags:
      - Materials
  /materials/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Upload image for material (admin only)
      parameters:
      - description: Material ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Image uploaded successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Upload material image
      tags:
      - Materials
  /test-upload:
    post:
      consumes:
      - multipart/form-data
      description: Test endpoint for file upload debugging
      parameters:
      - description: Test file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Test successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      summary: Test file upload
      tags:
      - Debug
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Используется для запросов через Insomnia/Postman: "Bearer <JWT>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"